♻️ refactor : Corrigindo para os novos endereçamentos de classes
