+++ src/main/java/br/com/unicuritiba/projetoathus/application/services/AuthService.java
@@ -3,8 +3,8 @@ package br.com.unicuritiba.projetoathus.application.services;
 import br.com.unicuritiba.projetoathus.infrastructure.mails.VerificarCadastro;
 import br.com.unicuritiba.projetoathus.domain.models.Usuario;
 import br.com.unicuritiba.projetoathus.domain.repositories.UsuarioRepository;
-import br.com.unicuritiba.projetoathus.domain.dto.LoginRequestDTO;
-import br.com.unicuritiba.projetoathus.domain.dto.RegisterRequestDTO;
+import br.com.unicuritiba.projetoathus.domain.dto.auth.LoginRequestDTO;
+import br.com.unicuritiba.projetoathus.domain.dto.auth.RegisterRequestDTO;
 import br.com.unicuritiba.projetoathus.infrastructure.exceptions.*;
 import com.auth0.jwt.interfaces.DecodedJWT;
 import lombok.RequiredArgsConstructor;
